Added solution - LeetHub

+++ Kadane's Algorithm - GFG/kadanes-algorithm.java
@@ -32,9 +32,13 @@ class Main {
 
 
 class Solution{
+
+    // arr: input array
+    // n: size of array
+    //Function to find the sum of contiguous subarray with maximum sum.
     long maxSubarraySum(int arr[], int n){
         
-        long currsum=0;long maxsum=Long.MIN_VALUE;
+        long currsum=0;long maxsum=Integer.MIN_VALUE;
         for(int i=0;i<n;i++){
             currsum+=arr[i];
             if(currsum>maxsum){
